firebase middleware

+++ server/models/cartModel.js
@@ -2,20 +2,21 @@
 
 const db = require('../models/db');
 
-exports.getAllItems = async () => {
+exports.getAllItems = async (userId) => {
   try {
-    const query = 'SELECT * FROM cart';
-    const result = await db.query(query);
+    const query = 'SELECT * FROM cart WHERE user_id = $1';
+    const values = [userId];
+    const result = await db.query(query, values);
     return result.rows;
   } catch (error) {
     throw error;
   }
 };
 
-exports.getItemById = async (cartId) => {
+exports.getItemById = async (cartId, userId) => {
   try {
-    const query = 'SELECT * FROM cart WHERE cart_id = $1';
-    const values = [cartId];
+    const query = 'SELECT * FROM cart WHERE cart_id = $1 and user_id = $2';
+    const values = [cartId, userId];
     const result = await db.query(query, values);
     return result.rows[0];
   } catch (error) {
@@ -52,10 +53,10 @@ exports.updateCartItem = async (cartId, user_id, plan_id, trainer_id) => {
   }
 };
 
-exports.deleteCartItem = async (cartId) => {
+exports.deleteCartItem = async (cartId, user_id) => {
   try {
-    const query = 'DELETE FROM cart WHERE cart_id = $1 RETURNING *';
-    const values = [cartId];
+    const query = 'DELETE FROM cart WHERE cart_id = $1 and user_id = $2 RETURNING *';
+    const values = [cartId, user_id];
     const result = await db.query(query, values);
     return result.rows[0];
   } catch (error) {
